treewide: Add files to execute an ndk-build for scanmem.
jni/Android.mk: Add a ndk-build Makefile, which does not support
libreadline.
jni/Application.mk: Add a ndk-build Application Makefile.
This can be modified to build for specific architectures.
An example of a specific architecture is given in the comments of
this file.
scanmem_jni: Add an empty stub for Java interfaces.